+++ CHANGELOG.md
@@ -23,6 +23,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - metrics for crawler (#50)
 - async name resolution (#62)
 - periodic crawler state snapshots (#63)
+- parser improvements and html entity decoding (#66)
 
 ### Fixed
 - Fix copying body data from buffer (#24)